requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.34.73
botocore==1.34.73
Django==4.2.11
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.1
excel-base==1.0.4
isoweek==1.3.3
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing-extensions==4.10.0
tzlocal==5.2
urllib3==1.26.18
xlwt==1.3.0
